requirements.txt
click==8.0.1
colorama==0.4.4
Flask==2.0.1
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
kafka-python==2.0.2
MarkupSafe==2.0.1
psycopg2==2.9.1
typing-extensions==3.10.0.2
Werkzeug==2.0.1
zipp==3.6.0
